requirements.txt
fastapi
uvicorn
openai
pytesseract
pdf2image
python-multipart
google-cloud-aiplatform
vertexai
